permission

+++ app/Http/Controllers/BackEnd/News/NewsController.php
@@ -23,34 +23,42 @@ class NewsController extends Controller
 
     public function index()
     {
-        // if(Auth::user()->can(PermissionConstant::NEWS_PERMISSION_LIST)) {
-            $news = $this->news->paginate(1);
+        if(Auth::user()->can(PermissionConstant::NEWS_PERMISSION_LIST)) {
+            $news = $this->news->sortable()->paginate(5);
             return view('backend.admin.news.index',compact('news'));
-        // }
+        }
     }
 
     public function view(Request $request)
     {
+        if(Auth::user()->can(PermissionConstant::NEWS_PERMISSION_VIEW)) {
         $news = $this->news->find($request->id);
         return view('backend.admin.news.view',compact('news'));
+        }
     }
 
     public function delete(Request $request)
     {
+        if(Auth::user()->can(PermissionConstant::NEWS_PERMISSION_DELETE)) {
         $news = $this->news->find($request->id);
         $news->delete();
         return redirect()->route('backend.admin.news.show')->with('success', Lang::get('message.delete', ['model' => 'Tin tức']));
+        }
     }
 
     public function newsDelete(Request $request)
     {
+        if(Auth::user()->can(PermissionConstant::NEWS_PERMISSION_DELETE)) {
         $this->news->whereIn('id', explode(",", $request->ids))->delete();
         return response()->json(['success' => "Xóa thành công"]);
+        }
     }
 
     public function create()
     {
+        if(Auth::user()->can(PermissionConstant::NEWS_PERMISSION_CREATE)) {
         return view('backend.admin.news.create');
+        }
     }
 
     public function store(Request $request)
@@ -66,8 +74,10 @@ class NewsController extends Controller
 
     public function edit(Request $request)
     {
+        if(Auth::user()->can(PermissionConstant::NEWS_PERMISSION_EDIT)) {
         $news = $this->news->find($request->id);
         return view('backend.admin.news.edit',compact('news'));
+        }
     }
 
     public function update(Request $request)
@@ -79,7 +89,7 @@ class NewsController extends Controller
         }else{
             $image = uploadFile($request->image, 'News_image');
         }
-        $this->news->create([
+        $this->news->update([
             'id' => $request->id,
             'title' => $request->title,
             'image' => uploadFile($request->image, 'News_image'),
